Refactor CLI command handling and improve error reporting
