+++ .github/workflows/maven.yml
@@ -69,7 +69,7 @@ jobs:
             This is the release of Alohomora version ${{ github.ref_name }}.
             The JAR files are attached below. (client and background)
           files: |
-            target/Alohomora-*-client.jar
-            target/Alohomora-*-background.jar
+            backgroundServices/target/Alohomora-client-*.jar
+            clientApp/target/Alohomora-background-*.jar
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
