+++ .idea/encodings.xml
@@ -1,9 +1,15 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/ClientApp/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/ClientApp/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/backgroundServices/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/backgroundServices/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/clientApp/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/clientApp/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java/io/dayfit/github/backgroundServices/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java/io/dayfit/github/backgroundServices/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java/io/dayfit/github/clientApp/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java/io/dayfit/github/clientApp/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
   </component>
 </project>
