+++ backgroundServices/src/main/java/io/dayfit/github/backgroundServices/configurations/BackgroundApplicationConfiguration.java
@@ -1,0 +1,11 @@
+package io.dayfit.github.backgroundServices.configurations;
+
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * Configuration class for the background application.
+ */
+@Configuration
+public class BackgroundApplicationConfiguration {
+
+}
