+++ backgroundServices/src/main/java/io/dayfit/github/backgroundServices/managers/ShutdownManager.java
