+++ backgroundServices/src/main/java/io/dayfit/github/backgroundServices/services/CLICommandService.java
