+++ backgroundServices/src/main/resources/application.properties
@@ -1,0 +1,6 @@
+logging.level.root=WARN
+logging.level.org.springframework=WARN
+
+server.ping.response=@server.ping.response@
+spring.main.banner-mode=off
+server.port=8080
