+++ clientApp/src/main/java/io/dayfit/github/clientApp/PasswordManager.java
