+++ clientApp/src/main/resources/META-INF/MANIFEST.MF
@@ -1,0 +1,3 @@
+Manifest-Version: 0.2.0
+Main-Class: io.dayfit.github.clientApp.Application
+
