+++ clientApp/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+application.version=@application.version@
+max.wait.time.backgroundServices=@max.wait.time.backgroundServices@
+
+server.ping.response=@server.ping.response@
+server.port=8080
