+++ pom.xml
@@ -1,99 +1,23 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
+    <groupId>io.github.dayfit</groupId>
+    <artifactId>Alohomora</artifactId>
+    <version>0.2.0</version>
+    <packaging>pom</packaging>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
         <version>3.4.4</version>
-        <relativePath/> <!-- lookup parent from repository -->
+        <relativePath/>
     </parent>
 
-    <groupId>io.github.dayfit</groupId>
-    <artifactId>Alohomora</artifactId>
-    <version>0.2.0</version>
-    <packaging>jar</packaging>
+    <modules>
+        <module>backgroundServices</module>
+        <module>clientApp</module>
+    </modules>
 
     <build>
-        <plugins>
-            <plugin>
-                <groupId>org.springframework.boot</groupId>
-                <artifactId>spring-boot-maven-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>client-app</id>
-                        <goals>
-                            <goal>repackage</goal>
-                        </goals>
-                        <configuration>
-                            <mainClass>io.dayfit.github.clientApp.Application</mainClass>
-                            <classifier>client</classifier>
-                            <layout>JAR</layout>
-                            <requiresUnpack>
-                                <dependency>
-                                    <groupId>${project.groupId}</groupId>
-                                    <artifactId>${project.artifactId}</artifactId>
-                                </dependency>
-                            </requiresUnpack>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>background-service</id>
-                        <goals>
-                            <goal>repackage</goal>
-                        </goals>
-                        <configuration>
-                            <mainClass>io.dayfit.github.backgroundServices.BackgroundApplication</mainClass>
-                            <classifier>background</classifier>
-                            <layout>JAR</layout>
-                            <requiresUnpack>
-                                <dependency>
-                                    <groupId>${project.groupId}</groupId>
-                                    <artifactId>${project.artifactId}</artifactId>
-                                </dependency>
-                            </requiresUnpack>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <!-- Add Maven JAR plugin to create main JARs -->
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-jar-plugin</artifactId>
-                <version>3.3.0</version>
-                <executions>
-                    <execution>
-                        <id>client-jar</id>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>jar</goal>
-                        </goals>
-                        <configuration>
-                            <classifier>client-classes</classifier>
-                            <includes>
-                                <include>io/dayfit/github/clientApp/**</include>
-                                <include>io/dayfit/github/shared/**</include>
-                            </includes>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>background-jar</id>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>jar</goal>
-                        </goals>
-                        <configuration>
-                            <classifier>background-classes</classifier>
-                            <includes>
-                                <include>io/dayfit/github/backgroundServices/**</include>
-                                <include>io/dayfit/github/shared/**</include>
-                            </includes>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
         <resources>
             <resource>
                 <directory>src/main/resources</directory>
@@ -102,8 +26,16 @@
         </resources>
     </build>
 
+    <properties>
+        <maven.compiler.source>21</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+
+        <application.version>${project.version}</application.version>
+        <max.wait.time.backgroundServices>5000</max.wait.time.backgroundServices>
+    </properties>
+
     <dependencies>
-        <!-- Dependencies remain unchanged -->
         <!--TESTING DEPENDENCIES-->
         <dependency>
             <groupId>org.junit.jupiter</groupId>
@@ -131,39 +63,10 @@
             <version>2.18.3</version>
         </dependency>
 
-        <!--Spring-boot-->
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-test</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-web</artifactId>
-        </dependency>
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
             <scope>provided</scope>
         </dependency>
-
-        <dependency>
-            <groupId>ch.qos.logback</groupId>
-            <artifactId>logback-classic</artifactId>
-        </dependency>
     </dependencies>
-
-    <properties>
-        <server.ping.response>ALOHOMORA_SERVER</server.ping.response>
-        <max.wait.time.backgroundServices>30000</max.wait.time.backgroundServices>
-        <application.version>${project.version}</application.version>
-        <maven.compiler.source>21</maven.compiler.source>
-        <maven.compiler.target>21</maven.compiler.target>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    </properties>
 </project>
